updating the regex form validation
